feat: crud completo

+++ src/main/java/com/seidelsoft/ERPBackend/endereco/service/CidadeService.java
@@ -6,6 +6,7 @@ import com.seidelsoft.ERPBackend.system.service.BaseService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,22 +19,51 @@ public class CidadeService extends BaseService<Cidade> {
     CidadeRepository cidadeRepository;
 
     @Override
+    public boolean validar(Cidade entity) {
+        if (entity.getNome() == null || entity.getNome().isEmpty()) {
+            return false;
+        }
+        if (entity.getEstado() == null || entity.getEstado().getUf() == null || entity.getEstado().getUf().isEmpty()) {
+            return false;
+        }
+        if (entity.getIbge() == null || entity.getIbge().isEmpty()) {
+            return false;
+        }
+        
+        return true;
+    }
+
+    @Override
     public Optional<Cidade> getById(Long id) {
         return cidadeRepository.findById(id);
     }
 
     @Override
-    public Page<Cidade> findAll(Pageable pageable) {
+    public List<Cidade> findAll(Sort sort) {
+        if (sort != null) {
+            return cidadeRepository.findAll(sort);
+        }
+        return null;
+    }
+
+    @Override
+    public Page<Cidade> findAllPaged(Pageable pageable) {
         return cidadeRepository.findAll(pageable);
     }
 
     @Override
     public void save(Cidade entity) {
+        if (!validar(entity)) {
+            throw new IllegalArgumentException("Dados inválidos para a cidade");
+        }
         cidadeRepository.save(entity);
     }
 
     @Override
     public void delete(Long id) {
+        if (id == null || !cidadeRepository.existsById(id)) {
+            throw new IllegalArgumentException("Estado não encontrado");
+        }
         cidadeRepository.deleteById(id);
     }
